requirements.txt
discord.py ~= 1.6.0
sqlitedict ~= 1.7.0
tabulate  ~= 0.8.9
python-dotenv